+++ src/app/_components/invitation/invitationPage.tsx
@@ -3,7 +3,6 @@ import { View, Heading } from "@adobe/react-spectrum";
 import { Suspense } from "react";
 import { Loading } from "../Loading";
 import { InvitationInfo } from "./invitationInfo";
-import { useSetAtom } from "jotai";
 import { useHydrateAtoms } from "jotai/utils";
 import { friendRequestTokenAtom } from "~/store";
