+++ src/app/friend/page.tsx
@@ -1,6 +1,6 @@
 import { FriendListPage } from "~/app/_components/friend";
 import { api, HydrateClient } from "~/trpc/server";
-import { Metadata } from "next";
+import { type Metadata } from "next";
 
 export const metadata: Metadata = {
   title: "友達一覧",
