+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { api, HydrateClient } from "~/trpc/server";
 import { createClient } from "~/utils/supabase/server";
-import { Metadata } from "next";
+import { type Metadata } from "next";
 
 export const metadata: Metadata = { title: "top" };
